Changes to apply brake deadband properly (according to some slack comments, like here: https://carnd.slack.com/archives/C6NVDVAQ3/p1505233214000427?thread_ts=1505227052.000090&cid=C6NVDVAQ3  )
Basically the idea is avoid applying brake to take advantage of the inherent brake of the car when no throttle is applied.
This avoids flipping quickly between throttle and brake.

The results in the 1st red light from start is a reduction of 50% aprox of the times the brake is pushed (tested at 10Hz)